WIP: Moving to short indicies

+++ engines/twp/gfx.h
@@ -175,10 +175,10 @@ public:
 
 	void clear(Color color);
 	void drawPrimitives(uint32 primitivesType, Vertex *vertices, int v_size, Math::Matrix4 transf = Math::Matrix4(), Texture *texture = NULL);
-	void drawPrimitives(uint32 primitivesType, Vertex *vertices, int v_size, uint32 *indices, int i_size, Math::Matrix4 transf = Math::Matrix4(), Texture *texture = NULL);
+	void drawPrimitives(uint32 primitivesType, Vertex *vertices, int v_size, unsigned short *indices, int i_size, Math::Matrix4 transf = Math::Matrix4(), Texture *texture = NULL);
 	void drawLines(Vertex *vertices, int count, Math::Matrix4 trsf = Math::Matrix4());
 	void drawLinesLoop(Vertex *vertices, int count, Math::Matrix4 trsf = Math::Matrix4());
-	void draw(Vertex *vertices, int v_size, uint32 *indices, int i_size, Math::Matrix4 trsf = Math::Matrix4(), Texture *texture = NULL);
+	void draw(Vertex *vertices, int v_size, unsigned short *indices, int i_size, Math::Matrix4 trsf = Math::Matrix4(), Texture *texture = NULL);
 	void drawQuad(Math::Vector2d size, Color color = Color(), Math::Matrix4 trsf = Math::Matrix4());
 	void drawSprite(Common::Rect textRect, Texture &texture, Color color = Color(), Math::Matrix4 trsf = Math::Matrix4(), bool flipX = false, bool flipY = false);
 	void drawSprite(Texture &texture, Color color = Color(), Math::Matrix4 trsf = Math::Matrix4(), bool flipX = false, bool flipY = false);
